# swagger.json
{"openapi":"3.0.0","paths":{"/api/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/items/owned":{"get":{"operationId":"ItemsController_getOwnedItems","parameters":[],"responses":{"401":{"description":"Unauthorized user"},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemsGetOwnedItemsResponse"}}}}},"tags":["items"],"security":[{"bearer":[]}]}},"/api/v1/oauth/me":{"get":{"operationId":"OauthController_getMe","summary":"Returns logged user data","parameters":[],"responses":{"200":{"description":"Returns logged user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OauthGetMeResponse"}}}},"401":{"description":"Unauthorized user"}},"tags":["oauth"],"security":[{"bearer":[]}]}},"/api/v1/offers":{"get":{"operationId":"OffersController_findAll","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OffersFindAllResponse"}}}}},"tags":["offers"]}},"/api/v1/offers/{id}/purchase":{"post":{"operationId":"OffersController_purchase","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"401":{"description":"Unauthorized user"},"403":{"description":"User is not allowed to purchase this offer. Insufficient funds"},"404":{"description":"Offer not found"},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OffersPurchaseResponse"}}}}},"tags":["offers"],"security":[{"bearer":[]}]}},"/api/v1/wallets/balance":{"get":{"operationId":"WalletsController_getBalance","parameters":[],"responses":{"401":{"description":"Unauthorized user"},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WalletsGetBalanceResponse"}}}}},"tags":["wallets"],"security":[{"bearer":[]}]}}},"info":{"title":"gs-shop-api-be","description":"GameSwift Shop API - env: LOCAL","version":"0.0.1","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3005/"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"OwnedItemResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"}},"required":["id","name"]},"ItemsGetOwnedItemsResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/OwnedItemResponse"}}},"required":["items"]},"OauthGetMeResponse":{"type":"object","properties":{"id":{"type":"string"},"nickname":{"type":"string"},"email":{"type":"string"},"avatarUrl":{"type":"string","nullable":true},"country":{"type":"string","nullable":true},"ranking":{"type":"number"}},"required":["id","nickname","email","avatarUrl","country","ranking"]},"OfferResponse":{"type":"object","properties":{"id":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"}},"required":["id","description","price"]},"OffersFindAllResponse":{"type":"object","properties":{"offers":{"type":"array","items":{"$ref":"#/components/schemas/OfferResponse"}}},"required":["offers"]},"OffersPurchaseResponse":{"type":"object","properties":{"orderId":{"type":"string"}},"required":["orderId"]},"TokenResponse":{"type":"object","properties":{"symbol":{"type":"string"},"icon":{"type":"string","description":"URL of the token icon"}},"required":["symbol","icon"]},"WalletsApiBalanceResponse":{"type":"object","properties":{"balance":{"type":"number"},"token":{"$ref":"#/components/schemas/TokenResponse"}},"required":["balance","token"]},"WalletsGetBalanceResponse":{"type":"object","properties":{"balances":{"type":"array","items":{"$ref":"#/components/schemas/WalletsApiBalanceResponse"}}},"required":["balances"]}}}}